# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Tracker Lite API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type TruckStatus = typeof TruckStatus[keyof typeof TruckStatus];


export const TruckStatus = {
  available: 'available',
  assigned: 'assigned',
  running: 'running',
  maintenance: 'maintenance',
} as const;

export interface Truck {
  id: number;
  registrationNumber: string;
  model: string;
  capacity: number;
  fuelType: string;
  status: TruckStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type TruckInputFuelType = typeof TruckInputFuelType[keyof typeof TruckInputFuelType];


export const TruckInputFuelType = {
  diesel: 'diesel',
  petrol: 'petrol',
  electric: 'electric',
  hybrid: 'hybrid',
} as const;

export interface TruckInput {
  /** @minLength 1 */
  registrationNumber: string;
  /** @minLength 1 */
  model: string;
  capacity: number;
  fuelType: TruckInputFuelType;
  notes?: string;
}

export type TruckUpdateFuelType = typeof TruckUpdateFuelType[keyof typeof TruckUpdateFuelType];


export const TruckUpdateFuelType = {
  diesel: 'diesel',
  petrol: 'petrol',
  electric: 'electric',
  hybrid: 'hybrid',
} as const;

export type TruckUpdateStatus = typeof TruckUpdateStatus[keyof typeof TruckUpdateStatus];


export const TruckUpdateStatus = {
  available: 'available',
  assigned: 'assigned',
  running: 'running',
  maintenance: 'maintenance',
} as const;

export interface TruckUpdate {
  /** @minLength 1 */
  registrationNumber?: string;
  /** @minLength 1 */
  model?: string;
  capacity?: number;
  fuelType?: TruckUpdateFuelType;
  status?: TruckUpdateStatus;
  notes?: string;
}

export interface Driver {
  id: number;
  name: string;
  licenseNumber: string;
  phone: string;
  address: string;
  experienceYears: number;
  isAvailable: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface DriverInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  licenseNumber: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @minimum 0 */
  experienceYears: number;
}

export interface DriverUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  licenseNumber?: string;
  /** @minLength 1 */
  phone?: string;
  /** @minLength 1 */
  address?: string;
  /** @minimum 0 */
  experienceYears?: number;
  isAvailable?: boolean;
}

export type TripStatus = typeof TripStatus[keyof typeof TripStatus];


export const TripStatus = {
  pending: 'pending',
  started: 'started',
  running: 'running',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Trip {
  id: number;
  pickupLocation: string;
  /** @nullable */
  pickupLat?: number | null;
  /** @nullable */
  pickupLng?: number | null;
  destination: string;
  /** @nullable */
  destLat?: number | null;
  /** @nullable */
  destLng?: number | null;
  /** @nullable */
  truckId?: number | null;
  /** @nullable */
  driverId?: number | null;
  /** @nullable */
  truckRegistration?: string | null;
  /** @nullable */
  driverName?: string | null;
  status: TripStatus;
  /** @nullable */
  distanceCovered?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface TripInput {
  /** @minLength 1 */
  pickupLocation: string;
  pickupLat?: number;
  pickupLng?: number;
  /** @minLength 1 */
  destination: string;
  destLat?: number;
  destLng?: number;
  truckId?: number;
  driverId?: number;
  notes?: string;
}

export type TripUpdateStatus = typeof TripUpdateStatus[keyof typeof TripUpdateStatus];


export const TripUpdateStatus = {
  pending: 'pending',
  started: 'started',
  running: 'running',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface TripUpdate {
  /** @minLength 1 */
  pickupLocation?: string;
  pickupLat?: number;
  pickupLng?: number;
  /** @minLength 1 */
  destination?: string;
  destLat?: number;
  destLng?: number;
  /** @nullable */
  truckId?: number | null;
  /** @nullable */
  driverId?: number | null;
  status?: TripUpdateStatus;
  notes?: string;
}

export interface TripLocation {
  id: number;
  tripId: number;
  lat: number;
  lng: number;
  /** @nullable */
  speed?: number | null;
  recordedAt: string;
}

export interface LocationInput {
  lat: number;
  lng: number;
  speed?: number;
}

export interface LiveTruckPosition {
  tripId: number;
  truckId: number;
  truckRegistration: string;
  driverName: string;
  lat: number;
  lng: number;
  /** @nullable */
  speed?: number | null;
  status: string;
  recordedAt: string;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  trip_started: 'trip_started',
  trip_completed: 'trip_completed',
  truck_delayed: 'truck_delayed',
  overspeed: 'overspeed',
  truck_offline: 'truck_offline',
} as const;

export interface Alert {
  id: number;
  type: AlertType;
  message: string;
  /** @nullable */
  tripId?: number | null;
  /** @nullable */
  truckId?: number | null;
  isRead: boolean;
  createdAt: string;
}

export type DashboardStatsMonthlyTripsItem = {
  month: string;
  count: number;
};

export type DashboardStatsTruckUtilizationItem = {
  status: string;
  count: number;
};

export interface DashboardStats {
  totalTrucks: number;
  availableTrucks: number;
  runningTrips: number;
  completedTrips: number;
  activeDrivers: number;
  delayedTrips: number;
  unreadAlerts?: number;
  monthlyTrips: DashboardStatsMonthlyTripsItem[];
  truckUtilization: DashboardStatsTruckUtilizationItem[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListTrucksParams = {
status?: string;
};

export type ListDriversParams = {
available?: boolean;
};

export type ListTripsParams = {
status?: string;
driverId?: number;
truckId?: number;
};

export type ListAlertsParams = {
unreadOnly?: boolean;
};