# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fleet Tracker Lite API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List all drivers
      parameters:
        - name: available
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Add a new driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
    put:
      operationId: updateDriver
      tags: [drivers]
      summary: Update a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Delete a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: driverId
          in: query
          required: false
          schema:
            type: integer
        - name: truckId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
    put:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{id}/start:
    post:
      operationId: startTrip
      tags: [trips]
      summary: Start a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
  /trips/{id}/complete:
    post:
      operationId: completeTrip
      tags: [trips]
      summary: Mark a trip as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
  /trips/{id}/locations:
    get:
      operationId: getTripLocations
      tags: [tracking]
      summary: Get location history for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripLocation"
    post:
      operationId: addTripLocation
      tags: [tracking]
      summary: Add a GPS location point (simulation)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Location recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripLocation"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/read:
    patch:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark an alert as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
components:
...
  schemas:
...
    Driver:
      type: object
      required: [id, name, licenseNumber, phone, address, experienceYears, isAvailable, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        licenseNumber:
          type: string
        phone:
          type: string
        address:
          type: string
        experienceYears:
          type: integer
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
    DriverInput:
      type: object
      required: [name, licenseNumber, phone, address, experienceYears]
      properties:
        name:
          type: string
          minLength: 1
        licenseNumber:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        experienceYears:
          type: integer
          minimum: 0
    DriverUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        licenseNumber:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        experienceYears:
          type: integer
          minimum: 0
        isAvailable:
          type: boolean
    Trip:
      type: object
      required: [id, pickupLocation, destination, status, createdAt]
      properties:
        id:
          type: integer
        pickupLocation:
          type: string
        pickupLat:
          type: ["number", "null"]
        pickupLng:
          type: ["number", "null"]
        destination:
          type: string
        destLat:
          type: ["number", "null"]
        destLng:
          type: ["number", "null"]
        truckId:
          type: ["integer", "null"]
        driverId:
          type: ["integer", "null"]
        truckRegistration:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, started, running, completed, cancelled]
        distanceCovered:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    TripInput:
      type: object
      required: [pickupLocation, destination]
      properties:
        pickupLocation:
          type: string
          minLength: 1
        pickupLat:
          type: number
        pickupLng:
          type: number
        destination:
          type: string
          minLength: 1
        destLat:
          type: number
        destLng:
          type: number
        truckId:
          type: integer
        driverId:
          type: integer
        notes:
          type: string
    TripUpdate:
      type: object
      properties:
        pickupLocation:
          type: string
          minLength: 1
        pickupLat:
          type: number
        pickupLng:
          type: number
        destination:
          type: string
          minLength: 1
        destLat:
          type: number
        destLng:
          type: number
        truckId:
          type: ["integer", "null"]
        driverId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, started, running, completed, cancelled]
        notes:
          type: string
    TripLocation:
      type: object
      required: [id, tripId, lat, lng, recordedAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        lat:
          type: number
        lng:
          type: number
        speed:
          type: ["number", "null"]
        recordedAt:
          type: string
          format: date-time
    LocationInput:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
        speed:
          type: number
...
    Alert:
      type: object
      required: [id, type, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [trip_started, trip_completed, truck_delayed, overspeed, truck_offline]
        message:
          type: string
        tripId:
          type: ["integer", "null"]
        truckId:
          type: ["integer", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
...