# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Tracker Lite API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlertType } from './alertType';

export interface Alert {
  id: number;
  type: AlertType;
  message: string;
  /** @nullable */
  tripId?: number | null;
  /** @nullable */
  truckId?: number | null;
  isRead: boolean;
  createdAt: Date;
}