# lib/api-zod/src/generated/types/tripLocation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Tracker Lite API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TripLocation {
  id: number;
  tripId: number;
  lat: number;
  lng: number;
  /** @nullable */
  speed?: number | null;
  recordedAt: Date;
}